+++ neoforge/src/main/java/net/mehvahdjukaar/supplementaries/integration/neoforge/DecoBlocksCompatImpl.java
@@ -58,7 +58,7 @@ public class DecoBlocksCompatImpl {
                         .noOcclusion()
                         .lightLevel((state) -> 11), CompatObjects.SOUL_CHANDELIER, () -> ParticleTypes.SOUL_FIRE_FLAME));
 
-        if (CompatHandler.DECO_BLOCKS_ABNORMALS) {
+        if (CompatObjects.ENDER_CHANDELIER.isPresent()) {
             ENDER_CHANDELIER_ROPE = RegHelper.registerBlock(Supplementaries.res("rope_ender_chandelier"), () ->
                     new RopeChandelierBlock(BlockBehaviour.Properties.of()
                             .strength(0.3F)
@@ -69,7 +69,7 @@ public class DecoBlocksCompatImpl {
                             CompatObjects.ENDER_FLAME));
         } else ENDER_CHANDELIER_ROPE = null;
 
-        if (CompatHandler.MUCH_MORE_MOD_COMPAT) {
+        if (CompatObjects.GLOW_CHANDELIER.isPresent()) {
             GLOW_CHANDELIER_ROPE = RegHelper.registerBlock(Supplementaries.res("rope_glow_chandelier"), () ->
                     new RopeChandelierBlock(BlockBehaviour.Properties.of()
                             .strength(0.3F)